backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
pydantic==2.12.2
pydantic-settings==2.11.0
boto3==1.40.52
mistralai==1.9.11
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters >=0.3.9, <1.0.0
langchain-aws==0.2.35
langchain-pinecone==0.2.12
pinecone==7.3.0
python-multipart==0.0.20
langgraph==1.0.0
perplexityai==0.17.0
tavily-python==0.7.12
